deskpro-backend/requirements.txt
Django>=5.0,<6.0
django-ninja>=1.3.0
psycopg[binary]>=3.1
djangorestframework-simplejwt>=5.3
python-decouple>=3.8
httpx>=0.27
django-cors-headers>=4.3